% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_analysis.R
\name{DEG_analysis}
\alias{DEG_analysis}
\title{DEG_analysis}
\usage{
DEG_analysis(
  filtered_count_table,
  sample_table,
  fdr_threshold = 0.05,
  logFC_threshold = 1
)
}
\arguments{
\item{filtered_count_table}{Count-table without low-counts filtered by CPM}

\item{sample_table}{computer path to the sample_table}

\item{fdr_threshold}{False discovery rate used for filtering DEGs}

\item{logFC_threshold}{Log2 fold change used for filtering DEGs}
}
\value{
A data frame of filtered DEGs
}
\description{
Create a table containing differential expressed genes
(using the filtered count_table) that
are filtered according to Log2 fold change and FDR
}
\examples{
# example code
sample_table_path = system.file("extdata","E-MTAB-2523_sample_table.txt",package = "RNAenrichmentanalysis")
 DEG_analysis =
 function (filtered_count_table,
 sample_table = sample_table_path,
 fdr_threshold = 0.05,
 logFC_threshold = 1 )


}
