% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_filter_data.R
\name{import_and_filter_data}
\alias{import_and_filter_data}
\title{import_and_filter_data}
\usage{
import_and_filter_data(count_table, min_CPM = 1)
}
\arguments{
\item{count_table}{computer path to the count_table}

\item{min_CPM}{Minimum threshold for filtering}
}
\value{
Filtered count table
}
\description{
Imports count_table using the file path
then filters the low expressed genes in the count_table
based on CPM
}
\examples{
 Filtered_count_table = import_and_filter_data(
 count_table = "counts_table.txt",
 min_CPM = 1
)

}
