% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Over_representation.R
\name{Over_representation_analysis}
\alias{Over_representation_analysis}
\title{Over_representation}
\usage{
Over_representation_analysis(filtered_DEG, GO = "ALL")
}
\arguments{
\item{filtered_DEG}{data.frame that will be used for the analysis}

\item{GO}{which of the 3 types of GO terms BP for biological pathways
MF for molecular function, CC for cellular component, otherwise ALL is default}
}
\value{
data results from the Over representation analysis named OR_data
}
\description{
use the analysed data from DEG_analysis (filtered_DEG)
convert EntrezIDs to SYMBOL
do an over representation analysis using clusterProfiler
All this requires the data from the earlier DEG_analysis
}
\examples{
Over_representation_analysis(filtered_DEG, GO="ALL")
to call upon GO data OR_data$GO for KEGG OR_data$KEGG


}
