% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{Plots}
\alias{Plots}
\title{Plots}
\usage{
Plots(OR_data, categories = 20)
}
\arguments{
\item{OR_data}{a list of the GO data and KEGG data for the graphs
ensure you use the $ to specify results of the enrichment}

\item{categories}{is how many categories would be used for the plot}
}
\value{
returns a list with the 3 plots
}
\description{
take the data from Over_representation (OR_data) and define category number
you want the plot to work with.
create 3 different plots based on the data
suppress warning messages
return as a list the three plots
}
