% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_filter_data.R
\name{import_and_filter_data}
\alias{import_and_filter_data}
\title{import_and_filter_data}
\usage{
import_and_filter_data(count_table, min_CPM = 1, min_sample = 2)
}
\arguments{
\item{min_CPM}{Minimum threshold for filtering}

\item{min_sample}{Minimum samples that pass the cpm threshold for every gene}

\item{count_table_path}{computer path to the count_table}
}
\value{
Filtered count table
}
\description{
Imports count_table using the file path
then filters the low expressed genes in the count_table
based on per sample CPM
}
\examples{
count_table_path = system.file("extdata","E-MTAB-2523.counts.txt",package = "RNAenrichmentanalysis")
filtered_count_table = import_and_filter_data(
count_table = count_table_path,
min_CPM = 1,
min_sample = 2
)


}
